Change AltoRoutes filename
GitHub Pages is case-sensitive for URL file paths. AltoRoutes logo file had a uppercase extension. Changing only the file extension to lowercase was no good, because git did not tracked the changes. So I had to rename the file for something which makes even more sense (the black ground in this case is black, not transparent).