Using HTTPS whenever possible
This is 2016, and we need to have HTTPS whenever we can. Unfortunatelly, some of our customers does not have SSL or a certificate, so we cannot do it.

Nevertheless, the main problem was loading Google scripts from non-HTTPS sources, which was perceived as a security flaw for Chrome and other browsers